+++ main.c
@@ -215,11 +215,11 @@ static const char *read_from_key_file(const char *filename) {
     }
 
     fseek(file, 0, SEEK_END);
-    unsigned file_size = ftell(file);
+    long file_size = ftell(file);
     rewind(file);
 
     char *out = (char *)malloc((file_size + 1) * sizeof(char));
-    unsigned end = fread(out, sizeof(char), file_size, file);
+    size_t end = fread(out, sizeof(char), file_size, file);
     out[end] = '\0';
 
     fclose(file);
